moved second and vote to seperate modules
